neo4j_data_loader/requirements.txt
neo4j-driver
kafka-python
retry